Reproducer
Our current exception serializer has trouble serializing and de-serializing exceptions which delegate serialization to another class.
If we fail to load the class which the serialization is delegated to, we immediately throw an exception during de-serialization, failing to continue to
read the message stream. Then, we try to keep reading the stream even though we haven't consumed the entire message. We should read the entire message
even if we fail to de-serialize the exception. Otherwise, we enter a state where we try to consume the remaining parts of the message as a new message.
This exposes an issue where if we DO fail to de-serialize a message, we do not fail the test immediately. Further, we also do not fail the test when
reaching an 'Unexpected exception thrown.' when the remaining message is interpreted as a new message.