Extend NamedDomainObjectContainerScope instead of re-defining String invoke extensions
This results in a binary breaking change due to the previous private constructor generating a public synthetic constructor
with an additional DefaultConstructorMarker parameter. This was then called by the companion object. However, it could not
be compiled against from Kotlin code since the constructor was private, so we accept the binary breaking change