Remove checking for duplicate test suite types
The existing behavior relied on attributes for all consumable configurations being realized. This is not the case anymore.
In reality, this sort of checking is done during variant metadata creation, though this is only done when a project
is actually consumed.

In general, test suites should have the same attributes, but with separate capabilities. We should get rid of the test suite
type attribute altogether.